move userQuery to query folder
