+++ components/User/SignIn.js
@@ -6,7 +6,7 @@ import { FaEye, FaEyeSlash } from 'react-icons/fa';
 import { useState } from 'react';
 import Form from '../styles/Form';
 import useForm from '../../lib/useForm';
-import { CURRENT_USER_QUERY } from '.';
+import { CURRENT_USER_QUERY } from '../../queries/getUser';
 import Error from '../ErrorMessage';
 import { ButtonStyles } from '../styles/StateStyles';
